Translatetd padDiff to ts.
